Replaced uint32_t with standard uint16_t for sequence_number variables.
R=stefan@webrtc.org

Review URL: https://codereview.webrtc.org/1232373004 .

Cr-Commit-Position: refs/heads/master@{#9588}

+++ webrtc/modules/remote_bitrate_estimator/test/bwe_test_framework.cc
@@ -175,7 +175,7 @@ MediaPacket::MediaPacket(int flow_id,
     : Packet(flow_id, send_time_us, payload_size), header_(header) {
 }
 
-MediaPacket::MediaPacket(int64_t send_time_us, uint32_t sequence_number)
+MediaPacket::MediaPacket(int64_t send_time_us, uint16_t sequence_number)
     : Packet(0, send_time_us, 0) {
   header_ = RTPHeader();
   header_.sequenceNumber = sequence_number;
